Avoid user input in ci or non-iteractive mode

+++ cmd/kyma/alpha/deploy/cmd.go
@@ -114,7 +114,7 @@ func (cmd *command) Run() error {
 		}
 
 		// delete workspace folder
-		if approvalRequired {
+		if approvalRequired && !cmd.avoidUserInteraction() {
 			userApprovalStep := stepFactory.AddStep("Workspace folder exists")
 			if userApprovalStep.PromptYesNo(fmt.Sprintf("Delete workspace folder '%s' after Kyma deployment?", cmd.opts.WorkspacePath)) {
 				defer os.RemoveAll(cmd.opts.WorkspacePath)
@@ -162,7 +162,7 @@ func (cmd *command) isCompatibleVersion() error {
 
 	//seemless upgrade unnecessary or cannot be warrantied - aks user for approval
 	qUpgradeIncompStep := stepFactory.AddStep("Continue Kyma upgrade")
-	if qUpgradeIncompStep.PromptYesNo("Do you want to proceed the upgrade? ") {
+	if cmd.avoidUserInteraction() || qUpgradeIncompStep.PromptYesNo("Do you want to proceed the upgrade? ") {
 		qUpgradeIncompStep.Success()
 		return nil
 	}
@@ -409,4 +409,9 @@ func (cmd *command) adminPw() (string, error) {
 		return "", err
 	}
 	return string(secret.Data["password"]), nil
+}
+
+//avoidUserInteraction returns true if user won't provide input
+func (cmd *command) avoidUserInteraction() bool {
+	return cmd.NonInteractive || cmd.CI
 }
