+++ cmd/kyma/alpha/uninstall/cmd.go
@@ -107,7 +107,7 @@ func (cmd *command) Run() error {
 			return err
 		}
 		// delete workspace folder
-		if approvalRequired {
+		if approvalRequired && !cmd.avoidUserInteraction() {
 			userApprovalStep := stepFactory.AddStep("Workspace folder exists")
 			if userApprovalStep.PromptYesNo(fmt.Sprintf("Delete workspace folder '%s' after Kyma was removed?", cmd.opts.WorkspacePath)) {
 				defer os.RemoveAll(cmd.opts.WorkspacePath)
@@ -190,4 +190,9 @@ func (cmd *command) retrieveKymaMetadata() (*metadata.KymaMetadata, error) {
 func (cmd *command) showSuccessMessage() {
 	// TODO: show processing summary
 	fmt.Println("Kyma successfully removed.")
+}
+
+//avoidUserInteraction returns true if user won't provide input
+func (cmd *command) avoidUserInteraction() bool {
+	return cmd.NonInteractive || cmd.CI
 }
